Fix style issues in Protocol project

+++ Protocol/Packets/Requests/CyclesPacket.cs
@@ -4,18 +4,15 @@
     {
         public ulong Cycles
         {
-            get => 
-                (ulong)
-                (
-                    (Data[5] << 0) | 
-                    (Data[6] << 8) | 
-                    (Data[7] << 16) | 
-                    (Data[8] << 24) |
-                    (Data[9] << 32) |
-                    (Data[10] << 40) |
-                    (Data[11] << 48) |
-                    (Data[12] << 56)
-                );
+            get =>
+                (ulong) Data[5] << 0 |
+                (ulong) Data[6] << 8 |
+                (ulong) Data[7] << 16 |
+                (ulong) Data[8] << 24 |
+                (ulong) Data[9] << 32 |
+                (ulong) Data[10] << 40 |
+                (ulong) Data[11] << 48 |
+                (ulong) Data[12] << 56;
             set
             {
                 Data[5] = (byte)value;
